# fd-microservice/openapi_spec.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
                  "/api/product/{code}":  {
                                              "get":  {
                                                          "tags":  [
                                                                       "product-controller"
                                                                   ],
                                                          "operationId":  "getProduct",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "code",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "string"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "404":  {
                                                                                        "description":  "Not Found",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "400":  {
                                                                                        "description":  "Bad Request",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "500":  {
                                                                                        "description":  "Internal Server Error",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "403":  {
                                                                                        "description":  "Forbidden",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/Product"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      },
                                              "put":  {
                                                          "tags":  [
                                                                       "product-controller"
                                                                   ],
                                                          "operationId":  "updateProduct",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "code",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "string"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ProductRequest"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "404":  {
                                                                                        "description":  "Not Found",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "400":  {
                                                                                        "description":  "Bad Request",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "500":  {
                                                                                        "description":  "Internal Server Error",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "403":  {
                                                                                        "description":  "Forbidden",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    },
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                          },
                  "/api/product":  {
                                       "post":  {
                                                    "tags":  [
                                                                 "product-controller"
                                                             ],
                                                    "operationId":  "createProduct",
                                                    "requestBody":  {
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ProductRequest"
                                                                                                                            }
                                                                                                             }
                                                                                    },
                                                                        "required":  true
                                                                    },
                                                    "responses":  {
                                                                      "404":  {
                                                                                  "description":  "Not Found",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              },
                                                                      "400":  {
                                                                                  "description":  "Bad Request",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              },
                                                                      "500":  {
                                                                                  "description":  "Internal Server Error",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              },
                                                                      "403":  {
                                                                                  "description":  "Forbidden",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              },
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              }
                                                                  }
                                                }
                                   },
...
                  "/api/product/all":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "product-controller"
                                                                ],
                                                       "operationId":  "getAllProducts",
                                                       "responses":  {
                                                                         "404":  {
                                                                                     "description":  "Not Found",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 },
                                                                         "400":  {
                                                                                     "description":  "Bad Request",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 },
                                                                         "500":  {
                                                                                     "description":  "Internal Server Error",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 },
                                                                         "403":  {
                                                                                     "description":  "Forbidden",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 },
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "type":  "array",
                                                                                                                                "items":  {
                                                                                                                                              "$ref":  "#/components/schemas/Product"
                                                                                                                                          }
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   }
                                       },
...
              },
    "components":  {
                       "schemas":  {
                                       "ApiResponse":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "success":  {
                                                                                              "type":  "boolean"
                                                                                          },
                                                                              "message":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                              "data":  {
                                                                                           "type":  "object"
                                                                                       },
                                                                              "timestamp":  {
                                                                                                "type":  "string",
                                                                                                "format":  "date-time"
                                                                                            }
                                                                          }
                                                       },
...
                                   }
                   }
}